refine code test=develop
